# backend/internal/app/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Pterodactyl Billing API
  version: 0.1.0
  description: |
    Billing platform for Pterodactyl Panel with promo-code-only top-ups.
    Money-in payment gateways are intentionally disabled.
servers:
  - url: /
paths:
...
  /api/auth/register:
    post:
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterInput" }
      responses:
        "201": { description: Created }
  /api/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200": { description: OK }
...
  /api/tariffs/{id}/purchase:
    post:
      summary: Purchase a tariff (debits balance, provisions server)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "201": { description: Created }
        "400": { description: Insufficient balance or inactive tariff }
...
  /api/admin/tariffs:
    post:
      summary: Admin create tariff
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Created }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }